(refactor) cleaned up code structure. mv inbound handler to dev dir
